Rename __mocks__ to fixtures
