+++ packages/react-scripts/template/fixtures/store.js
